Added a position option that allows you to set the position (in seconds) where the thumbnail should be taken. Right now this only works for iOS, but the Javascript interaction will change for both iOS and Android to this:
window.PKVideoThumbnail.createThumbnail ( sourceVideoPath, targetThumbnailPath, position, success, failure );

Position is an integer, so if you pass it something like "2.6", it'll truncate to the integer value, which will be "2". If you need it rounded instead of truncated, do the rounding in Javascript first, then pass the rounded integer to the PKVideoThumbnail library.